feat: Move the metrics handler fixture
