+++ test-common/metrics-handler-fixture.ts
